+++ src/views/site/view.php
@@ -1,0 +1,16 @@
+<?php
+/**
+ * @see SiteController::actionView()
+ *
+ * @var View $this
+ * @var Entry $entry
+ */
+
+use app\models\Entry;
+use davidhirtz\yii2\cms\controllers\SiteController;
+use davidhirtz\yii2\cms\widgets\MetaTags;
+use davidhirtz\yii2\cms\widgets\Sections;
+use davidhirtz\yii2\skeleton\web\View;
+
+MetaTags::register(['model' => $entry]);
+echo Sections::widget(['entry' => $entry,]);